Add authorization for editing and deleting books

+++ app/controllers/author/detail.js
@@ -1,11 +1,16 @@
 import Controller from '@ember/controller';
+import { inject as service } from '@ember/service';
 
 export default Controller.extend({
-    actions: {
-        deleteAuthor(author) {
-            author.destroyRecord().then(() => {
-                this.transitionToRoute('author.index'); // could also route to author
-            });
-        }
-    }
+  can: service(),
+
+  actions: {
+    deleteAuthor(author) {
+      if (this.get('can').can('delete author', this.model)) {
+        author.destroyRecord().then(() => {
+          this.transitionToRoute('author.index'); // could also route to author
+        });
+      }
+    },
+  },
 });
